extract path related functions in a specific file
