+++ src/app/models/routing/content-route.ts
@@ -1,8 +1,3 @@
-import { ParamMap } from '@angular/router';
-import { UrlCommandParameters } from '../../utils/url';
-
-/* for url */
-export const pathParamName = 'p';
 
 type Id = string;
 
@@ -10,22 +5,6 @@ export interface ContentRoute {
   contentType: string,
   id: Id,
   path: Id[],
-}
-
-export function pathFromRouterParameters(params: ParamMap): string[]|null {
-  const pathAsString = params.get(pathParamName);
-  if (pathAsString === null) return null;
-  return pathFromParamValue(pathAsString);
-}
-
-export function pathFromParamValue(path: string): string[] {
-  return path === '' ? [] : path.split(',');
-}
-
-export function pathAsParameter(value: string[]): UrlCommandParameters {
-  const params: UrlCommandParameters = {};
-  params[pathParamName] = value;
-  return params;
 }
 
 /*
