+++ src/app/models/routing/item-route.ts
@@ -1,12 +1,13 @@
 import { ParamMap } from '@angular/router';
 import { defaultAttemptId } from '../../items/models/attempts';
 import { appConfig } from '../../utils/config';
-import { ContentRoute, pathAsParameter, pathFromRouterParameters } from './content-route';
+import { ContentRoute } from './content-route';
 import { ItemTypeCategory } from '../../items/models/item-type';
 import { isString } from '../../utils/type-checkers';
 import { UrlCommand } from '../../utils/url';
 import { arraysEqual } from '../../utils/array';
 import { AnswerId, AttemptId, ItemId } from '../ids';
+import { pathAsParameter, pathFromRouterParameters } from './path-parameter';
 
 // url parameter names
 export const activityPrefix = 'a';
